% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_decimal.R
\name{decimal}
\alias{decimal}
\title{`decimal` vector}
\usage{
decimal(x = numeric(), precision = 0L, scale = 2L, strict = FALSE)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
An S3 vector of class `decimal`.
}
\description{
This creates a double vector that represents percentages so when it is
printed, it is multiplied by 100 and suffixed with `%`.
}
