% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helpers}
\alias{helpers}
\alias{silent}
\alias{is_numchar}
\alias{is_numeric}
\alias{n_int_digits}
\alias{max_int_digits}
\title{Helpers}
\usage{
silent(...)

is_numchar(x, verbose = FALSE)

is_numeric(x)

n_int_digits(x, .fn = NULL, ...)

max_int_digits(x = numeric())
}
\description{
Helpers
}
